# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for toy rental service",
        "title": "ToyRental API",
        "contact": {
            "name": "Indrawansyah",
            "email": "indra@dev.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/toy": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy"
                ],
                "summary": "Mengambil semua data toy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Toy"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy"
                ],
                "summary": "Menambahkan mainan baru",
                "parameters": [
                    {
                        "description": "Data mainan baru",
                        "name": "toy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ToyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Toy"
                        }
                    }
                }
            }
        },
        "/toy/category": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy Category"
                ],
                "summary": "Mengambil semua data toy category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ToyCategory"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy Category"
                ],
                "summary": "Insert toy category",
                "parameters": [
                    {
                        "description": "Toy Category",
                        "name": "toy_category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ToyCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/toy/category/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy Category"
                ],
                "summary": "Mengambil data toy category berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ToyCategory"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy Category"
                ],
                "summary": "Update toy berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Toy Category",
                        "name": "toy_category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ToyCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Toy Category"
                ],
                "summary": "Delete toy berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
...
        "/toy/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy"
                ],
                "summary": "Mengambil data toy berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Toy"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy"
                ],
                "summary": "Update mainan berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data mainan yang diperbarui",
                        "name": "toy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ToyUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Toy",
                        "schema": {
                            "$ref": "#/definitions/entity.Toy"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Toy"
                ],
                "summary": "Menghapus mainan berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toy Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Toy"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Toy": {
            "type": "object",
            "properties": {
                "age_recommendation": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ToyCategory"
                    }
                },
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ToyImage"
                    }
                },
                "is_available": {
                    "type": "boolean"
                },
                "late_fee_per_day": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "primary_image": {
                    "type": "string"
                },
                "rental_price": {
                    "type": "number"
                },
                "replacement_price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "entity.ToyCategory": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.ToyImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                }
            }
        },
        "entity.ToyRequest": {
            "type": "object",
            "required": [
                "category_ids",
                "condition",
                "image_ids",
                "late_fee_per_day",
                "name",
                "rental_price",
                "replacement_price",
                "stock"
            ],
            "properties": {
                "age_recommendation": {
                    "type": "string"
                },
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_available": {
                    "type": "boolean"
                },
                "late_fee_per_day": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "primary_image_id": {
                    "type": "string"
                },
                "rental_price": {
                    "type": "number"
                },
                "replacement_price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "entity.ToyUpdateRequest": {
            "type": "object",
            "required": [
                "category_ids",
                "condition",
                "image_ids",
                "late_fee_per_day",
                "name",
                "rental_price",
                "replacement_price",
                "stock"
            ],
            "properties": {
                "age_recommendation": {
                    "type": "string"
                },
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_available": {
                    "type": "boolean"
                },
                "late_fee_per_day": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "primary_image_id": {
                    "type": "string"
                },
                "rental_price": {
                    "type": "number"
                },
                "replacement_price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "response.APISuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}